# contracts/product/v1/openapi.yaml
# ┌─────────────────────────────────────────────────────────────────────┐
# │  Echo Product API v1                                                │
# │  Owns evaluation resources, gate decisions, per-sample feedback,     │
# │  and Catalyst-compatible training-candidate export. Not engine state.│
# └─────────────────────────────────────────────────────────────────────┘
openapi: 3.1.2
x-cyrene-contract-profile: product-http-v1
info:
  title: Cyrene Echo Product API
  version: 1.0.0
  license: {name: Apache-2.0}
paths:
  /api/v1/evaluation-suites:
    post:
      operationId: createEvaluationSuite
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSuiteRequest'}
      responses:
        '201':
          description: EvaluationSuite created or replayed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvaluationSuite'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-suites/{suiteId}:
    get:
      operationId: getEvaluationSuite
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      responses:
        '200':
          description: EvaluationSuite
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvaluationSuite'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/judge-profiles:
    post:
      operationId: createJudgeProfile
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateJudgeProfileRequest'}
      responses:
        '201':
          description: JudgeProfile created or replayed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JudgeProfile'}
        '409': {$ref: '#/components/responses/Problem'}
  /api/v1/judge-profiles/{profileId}:
    get:
      operationId: getJudgeProfile
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '200':
          description: JudgeProfile
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JudgeProfile'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/session-artifacts:
    post:
      operationId: importSessions
      requestBody:
        required: true
        content:
          application/jsonl:
            schema: {type: string, format: binary, description: UTF-8 JSONL of session/model-output records}
      responses:
        '201':
          description: Immutable dataset ArtifactRef published to the local plane
          content:
            application/json:
              schema: {$ref: './generated/platform/artifact-ref.schema.json'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-runs:
    post:
      operationId: createEvaluationRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRunRequest'}
      responses:
        '201':
          description: Completed EvaluationRun; follow resultId and gateId
          content:
            application/json:
              schema: {$ref: './evaluation-run.schema.json'}
        '404': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-runs/{runId}:
    get:
      operationId: getEvaluationRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Persisted EvaluationRun
          content:
            application/json:
              schema: {$ref: './evaluation-run.schema.json'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-runs/{runId}/samples:
    get:
      operationId: listSamples
      parameters:
        - $ref: '#/components/parameters/RunId'
        - {name: only_passed, in: query, required: false, schema: {type: boolean}}
        - {name: only_annotated, in: query, required: false, schema: {type: boolean}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 1000, default: 200}}
        - {name: offset, in: query, required: false, schema: {type: integer, minimum: 0, default: 0}}
      responses:
        '200':
          description: Per-sample records for the run
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/SampleRecord'}}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-runs/{runId}/samples/{sampleIndex}:
    get:
      operationId: getSample
      parameters:
        - $ref: '#/components/parameters/RunId'
        - {name: sampleIndex, in: path, required: true, schema: {type: integer, minimum: 1}}
      responses:
        '200':
          description: One SampleRecord
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SampleRecord'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-runs/{runId}/annotations:
    post:
      operationId: annotateSample
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAnnotationRequest'}
      responses:
        '201':
          description: HumanAnnotation created or updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HumanAnnotation'}
        '404': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
    get:
      operationId: listAnnotations
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Human annotations for the run
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/HumanAnnotation'}}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-results/{resultId}:
    get:
      operationId: getEvaluationResult
      parameters:
        - name: resultId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Immutable EvaluationResult
          content:
            application/json:
              schema: {$ref: './evaluation-result.schema.json'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/gate-decisions/{gateId}:
    get:
      operationId: getGateDecision
      parameters:
        - name: gateId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Immutable GateDecision
          content:
            application/json:
              schema: {$ref: './gate-decision.schema.json'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/feedback-sets:
    post:
      operationId: createFeedbackSet
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateFeedbackSetRequest'}
      responses:
        '201':
          description: FeedbackSet created from explicitly selected samples
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FeedbackSet'}
        '404': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    get:
      operationId: listFeedbackSets
      parameters:
        - {name: runId, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Feedback sets, optionally filtered by run
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/FeedbackSet'}}
  /api/v1/feedback-sets/{feedbackSetId}:
    get:
      operationId: getFeedbackSet
      parameters:
        - $ref: '#/components/parameters/FeedbackSetId'
      responses:
        '200':
          description: FeedbackSet
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FeedbackSet'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/feedback-sets/{feedbackSetId}/export:
    post:
      operationId: exportFeedbackSet
      parameters:
        - $ref: '#/components/parameters/FeedbackSetId'
      responses:
        '200':
          description: Export outcome; handoff is PREPARED, never faked
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExportFeedbackSetResponse'}
        '404': {$ref: '#/components/responses/Problem'}
    get:
      operationId: downloadFeedbackSet
      parameters:
        - $ref: '#/components/parameters/FeedbackSetId'
      responses:
        '200':
          description: Catalyst-compatible JSONL of training candidates
          content:
            application/jsonl:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/evaluation-inputs:
    $ref: './lifecycle.openapi.yaml#/paths/~1api~1v1~1evaluation-inputs'
  /api/v1/evaluation-inputs/{input_id}:
    $ref: './lifecycle.openapi.yaml#/paths/~1api~1v1~1evaluation-inputs~1{input_id}'
  /api/v1/evaluation-inputs/{input_id}/samples:
    $ref: './lifecycle.openapi.yaml#/paths/~1api~1v1~1evaluation-inputs~1{input_id}~1samples'
  /api/v1/evaluation-inputs/{input_id}/actions/evaluate:
    $ref: './lifecycle.openapi.yaml#/paths/~1api~1v1~1evaluation-inputs~1{input_id}~1actions~1evaluate'
  /api/v1/feedback-sets/{feedback_id}/actions/send-to-catalyst:
    $ref: './lifecycle.openapi.yaml#/paths/~1api~1v1~1feedback-sets~1{feedback_id}~1actions~1send-to-catalyst'
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 1, maxLength: 200}
    SuiteId:
      name: suiteId
      in: path
      required: true
      schema: {type: string, format: uuid}
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: string, format: uuid}
    ProfileId:
      name: profileId
      in: path
      required: true
      schema: {type: string, format: uuid}
    FeedbackSetId:
      name: feedbackSetId
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: {$ref: './generated/common/problem-details.schema.json'}
  schemas:
    CreateSuiteRequest:
      type: object
      required: [name, evaluator, expectedField, actualField, threshold]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        evaluator: {enum: [exact_match.v1, llm_judge.v1], default: exact_match.v1}
        expectedField: {type: string, minLength: 1}
        actualField: {type: string, minLength: 1}
        threshold: {type: number, minimum: 0, maximum: 1}
        judgeProfileId: {type: string, format: uuid}
      additionalProperties: false
    EvaluationSuite:
      type: object
      required: [id, name, evaluator, expectedField, actualField, threshold, judgeProfileId, state, createdAt, updatedAt, resourceVersion]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        evaluator: {enum: [exact_match.v1, llm_judge.v1]}
        expectedField: {type: string}
        actualField: {type: string}
        threshold: {type: number, minimum: 0, maximum: 1}
        judgeProfileId: {type: string, format: uuid, nullable: true}
        state: {enum: [ACTIVE, ARCHIVED]}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resourceVersion: {type: integer, minimum: 1}
      additionalProperties: false
    CreateJudgeProfileRequest:
      type: object
      required: [name, judgeModel, exchangeEndpointRef, promptTemplate]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        judgeModel: {type: string, minLength: 1, maxLength: 200}
        exchangeEndpointRef: {type: string, minLength: 1, maxLength: 500}
        promptTemplate: {type: string, minLength: 1, maxLength: 4000}
      additionalProperties: false
    JudgeProfile:
      type: object
      required: [id, name, judgeModel, exchangeEndpointRef, promptTemplate, createdAt, updatedAt, resourceVersion]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        judgeModel: {type: string}
        exchangeEndpointRef: {type: string}
        promptTemplate: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resourceVersion: {type: integer, minimum: 1}
      additionalProperties: false
    CreateRunRequest:
      type: object
      required: [suiteId, inputArtifact, engineBindingId]
      properties:
        suiteId: {type: string, format: uuid}
        inputArtifact: {$ref: './generated/platform/artifact-ref.schema.json'}
        engineBindingId: {type: string, minLength: 1, maxLength: 200}
      additionalProperties: false
    SampleRecord:
      type: object
      required: [id, runId, sampleIndex, evaluator, inputDigest, inputRecord, passed, score, createdAt, resourceVersion]
      properties:
        id: {type: string, format: uuid}
        runId: {type: string, format: uuid}
        sampleIndex: {type: integer, minimum: 1}
        evaluator: {type: string}
        inputDigest: {type: string, pattern: "^sha256:[0-9a-f]{64}$"}
        inputRecord: {type: object}
        expected: {}
        actual: {}
        rawOutput: {type: string, nullable: true}
        passed: {type: boolean}
        score: {type: number, minimum: 0, maximum: 1}
        modelRef: {type: string, nullable: true}
        endpointRef: {type: string, nullable: true}
        usage: {$ref: '#/components/schemas/UsageFacts', nullable: true}
        judgeIdentity: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
        resourceVersion: {type: integer, minimum: 1}
      additionalProperties: false
    UsageFacts:
      type: object
      required: [source]
      properties:
        promptTokens: {type: integer, minimum: 0, nullable: true}
        completionTokens: {type: integer, minimum: 0, nullable: true}
        totalTokens: {type: integer, minimum: 0, nullable: true}
        source: {type: string}
      additionalProperties: false
    CreateAnnotationRequest:
      type: object
      required: [sampleIndex, reviewer]
      properties:
        sampleIndex: {type: integer, minimum: 1}
        reviewer: {type: string, minLength: 1, maxLength: 200}
        manualScore: {type: number, minimum: 0, maximum: 1, nullable: true}
        correctedOutput: {type: string, maxLength: 10000, nullable: true}
        preference: {enum: [accept, reject, prefer_corrected], nullable: true}
        note: {type: string, maxLength: 2000}
      additionalProperties: false
    HumanAnnotation:
      type: object
      required: [id, runId, sampleIndex, reviewer, note, createdAt, updatedAt, resourceVersion]
      properties:
        id: {type: string, format: uuid}
        runId: {type: string, format: uuid}
        sampleIndex: {type: integer, minimum: 1}
        reviewer: {type: string}
        manualScore: {type: number, minimum: 0, maximum: 1, nullable: true}
        correctedOutput: {type: string, nullable: true}
        preference: {enum: [accept, reject, prefer_corrected], nullable: true}
        note: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resourceVersion: {type: integer, minimum: 1}
      additionalProperties: false
    CreateFeedbackSetRequest:
      type: object
      required: [name, runId, sampleIndexes]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        runId: {type: string, format: uuid}
        sampleIndexes: {type: array, minItems: 1, items: {type: integer, minimum: 1}}
        annotationIds: {type: array, items: {type: string, format: uuid}}
      additionalProperties: false
    FeedbackSet:
      type: object
      required: [id, name, runId, sampleIndexes, annotationIds, state, handoffStatus, candidateCount, createdAt, updatedAt, resourceVersion]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        runId: {type: string, format: uuid}
        sampleIndexes: {type: array, items: {type: integer, minimum: 1}}
        annotationIds: {type: array, items: {type: string, format: uuid}}
        state: {enum: [OPEN, EXPORTED]}
        exportArtifact: {$ref: './generated/platform/artifact-ref.schema.json', nullable: true}
        handoffStatus: {enum: [PREPARED, HANDLED_OFF]}
        candidateCount: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resourceVersion: {type: integer, minimum: 1}
      additionalProperties: false
    ExportFeedbackSetResponse:
      type: object
      required: [feedbackSet, exportArtifact, candidateCount, handoffStatus]
      properties:
        feedbackSet: {$ref: '#/components/schemas/FeedbackSet'}
        exportArtifact: {$ref: './generated/platform/artifact-ref.schema.json'}
        candidateCount: {type: integer, minimum: 0}
        handoffStatus: {enum: [PREPARED, HANDLED_OFF]}
      additionalProperties: false